init home dir and db at startup. also, use public fields for api beans.

+++ fcrepo-cloudsync/fcrepo-cloudsync-api/src/main/java/com/github/cwilper/fcrepo/cloudsync/api/Configuration.java
@@ -5,7 +5,7 @@ import javax.xml.bind.annotation.XmlRootElement;
 @XmlRootElement(name="configuration")
 public class Configuration {
 
-    public Long keepSysLogDays;
-    public Long keepTaskLogDays;
+    public Integer keepSysLogDays;
+    public Integer keepTaskLogDays;
 
 }
